Rename folder DriftModifiers to GreenFunctionModifiers
