+++ src/QMCDrivers/GreenFunctionModifiers/DriftModifierBase.h
